backend/requirements.txt
django
djangorestframework
drf-yasg
requests~=2.32.3
ddddocr~=1.5.6
schedule~=1.2.2
django-cors-headers
pyexecjs2
psycopg2
scipy
pandas
GDAL==3.6.4
django-filter
xarray
joblib
numpy
geopandas
lightgbm
scikit-learn==1.4.2
